+++ src/common/ProfileInput.tsx
@@ -1,17 +1,15 @@
-export default function EditProfileInput() {
+export default function ProfileInput() {
   return (
     <div>
       <label htmlFor="nickname" className="text-[16px] text-left block font-semibold mt-7">
         닉네임
       </label>
-      <div className="">
-        <input
-          type="nickname"
-          id="nickname"
-          placeholder="현재 닉네임"
-          className="border-b-[1.4px] border-[#C1C7CD] text-[14px] text-[#697077] bg-[#F2F4F8] placeholder-[#697077] focus:outline-none p-2 w-full mt-3 h-[40px] rounded-md "
-        />
-      </div>
+      <input
+        type="nickname"
+        id="nickname"
+        placeholder="현재 닉네임"
+        className="border-b-[1.4px] border-[#C1C7CD] text-[14px] text-[#697077] bg-[#F2F4F8] placeholder-[#697077] focus:outline-none p-2 w-full mt-3 h-[40px] rounded-md "
+      />
       <label htmlFor="bio" className="text-[16px] text-left block font-semibold mt-5">
         한 줄 소개
       </label>
@@ -21,11 +19,6 @@ export default function EditProfileInput() {
         maxLength={50}
         className="border-b-[1.4px] border-[#C1C7CD] text-[14px] text-[#697077] bg-[#F2F4F8] placeholder-[#697077] focus:outline-none p-2 w-full mt-3 h-[63px] rounded-md resize-none"
       />
-      <div className="text-center mt-5 ">
-        <button className="text-[23px] font-semibold w-full bg-mainColor h-[63px] text-white p-2 rounded-md">
-          변경하기
-        </button>
-      </div>
     </div>
   );
 }
