+++ src/components/EditProfile/EditButton.tsx
@@ -1,0 +1,9 @@
+export default function EditButton() {
+  return (
+    <div className="text-center mt-5 ">
+      <button className="text-[23px] font-semibold w-full bg-mainColor h-[63px] text-white p-2 rounded-md">
+        변경하기
+      </button>
+    </div>
+  );
+}
