+++ src/components/Mypage/Footer.tsx
