+++ src/components/Mypage/LogoutBtn.tsx
@@ -1,3 +1,3 @@
 export default function LogoutBtn() {
-  return <div className="flex-1 text-[16px] font-semibold mt-[10px]">로그아웃</div>;
+  return <div className="flex-1 text-[16px] font-semibold mt-[10px] mb-[20px]">로그아웃</div>;
 }
