+++ src/pages/Profile.tsx
@@ -1,6 +1,8 @@
 import Header from '../common/Header';
+import ProfileImg from '../common/ProfileImg';
+import ProfileInput from '../common/ProfileInput';
 import Button from '../components/Profile/Button';
-import ProfileInput from '../components/Profile/ProfileInput';
+
 import Title from '../components/Profile/Title';
 
 export default function Profile() {
@@ -9,6 +11,7 @@ export default function Profile() {
       <Header title="" navigateTo="/emailverification" />
       <div className="px-[36px] flex-grow">
         <Title />
+        <ProfileImg />
         <ProfileInput />
       </div>
       <div className="absolute bottom-[48px] left-0 right-0 px-[30px]">
